create relative path
i have an alteryx file which has many excel sheets as inputs, but the file locations are not relative .they are absolute. i want to make this relative so that even if i move my file to another location i should be able to run the alteryx files without any issues.